Crate coz

Rust support for the coz Causal Profiler

This crate is a translation of the coz.h header file provided by coz to Rust, and enables profiling Rust programs with coz to profile throughput and latency.

For usage information, consult the README.md for the coz repository as well as the README.md for coz-rs.

Macros

begin
Equivalent of the COZ_BEGIN macro

end
Equivalent of the COZ_END macro

progress
Equivalent of the COZ_PROGRESS and COZ_PROGRESS_NAMED macros

Structs

Counter
A coz-counter which is either intended for throughput or begin/end points.

Functions

thread_init
Perform one-time per-thread initialization for coz.